# packages/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  login: string;
  password: string;
}