# packages/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  name: string;
  login: string;
  role: UserRole;
  status: UserStatus;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}